polish(cloud-iam): G004 slice 1 review pass — rustfmt, clippy, ADR D6.0 caller-authn
Hostile pre-open review findings applied: rustfmt across all new files,
clippy double_must_use/unnecessary_cast fixes, and ADR-0559 D6 gains
step 0 — PDP-API caller authentication (workload mTLS/SPIFFE) is a
named prerequisite before any consumer pointing, making the slice-1
trust boundary explicit.

+++ cloud/cloud-iam/crates/oya-cloud-iam-pdp-app/src/idgen.rs
@@ -53,7 +53,7 @@ impl IdGenerator for SystemUlidIdGenerator {
             .duration_since(std::time::UNIX_EPOCH)
             .map(|d| d.as_millis())
             .unwrap_or(0); // pre-epoch clock: encode as the epoch, never panic
-        let timestamp = (millis as u128) & ((1u128 << 48) - 1);
+        let timestamp = millis & ((1u128 << 48) - 1);
         let mut entropy = [0u8; 10];
         self.rng.fill(&mut entropy).map_err(|_| {
             // The kernel error enum has no entropy variant; the detail string
